update file api service and controller

+++ exzellenzkoch/src/main/kotlin/de/psekochbuch/exzellenzkoch/application/service/FileService.kt
@@ -1,13 +1,16 @@
 package de.psekochbuch.exzellenzkoch.application.service
 
+import com.google.firebase.auth.FirebaseAuth
 import de.psekochbuch.exzellenzkoch.application.dto.FileDto
 import de.psekochbuch.exzellenzkoch.domain.model.PublicRecipe
 import de.psekochbuch.exzellenzkoch.infrastructure.dao.PublicRecipeDao
+import de.psekochbuch.exzellenzkoch.security.firebase.FirebaseAuthentication
 import org.apache.commons.io.FilenameUtils
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.core.io.InputStreamResource
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.multipart.MultipartFile
@@ -39,16 +42,16 @@ class FileService {
      * Create folderlocation for the image. Then it creates the file. After that it creates the URL
      * which the image can load and at it to the recipe
      * @param file The file to save
-     * @param recipeId The id of the recipe, which should has the file
      * @return The URL from the image
      */
-    fun addImage(file: MultipartFile, recipeId:Int) :FileDto?
+    fun addImage(file: MultipartFile) :FileDto?
     {
-        val recipe : PublicRecipe = recipeDao?.findById(recipeId)?.get() ?: return null
+        val auth : FirebaseAuthentication = SecurityContextHolder.getContext().authentication as FirebaseAuthentication
+        val userId = FirebaseAuth.getInstance().getUser(auth.principal as String).displayName
         val fileName = file.originalFilename
 
         val name = FilenameUtils.removeExtension(file.originalFilename)
-        val folderPath = "images/" + recipe.user?.userId
+        val folderPath = "images/$userId"
         val folderFile = File(folderPath)
         if(!folderFile.exists()) folderFile.mkdir()
 
@@ -58,20 +61,17 @@ class FileService {
         if(!localFile.exists())
         {
             localFile.writeBytes(file.bytes)
-            recipe.picture = "api/$localPath"
-            recipeDao.save(recipe)
             return FileDto("api/$localPath")
         }
         var count = 1
-        var editFileName = "images/"+ recipe.user?.userId + "/$name" + "_$count" + FilenameUtils.getExtension(fileName)
+        var editFileName = "images/"+ userId + "/$name" + "_$count" + FilenameUtils.getExtension(fileName)
         while(File(editFileName).exists())
         {
             count++
-            editFileName = "images/"+ recipe.user?.userId + "/$name" + "_$count" + FilenameUtils.getExtension(fileName)
+            editFileName = "images/"+ userId + "/$name" + "_$count" + FilenameUtils.getExtension(fileName)
         }
         File(editFileName).writeBytes(file.bytes)
-        recipe.picture = "api/$editFileName"
-        recipeDao.save(recipe)
+
         return FileDto("api/$editFileName")
     }
 
@@ -81,18 +81,21 @@ class FileService {
      * @param recipeId The id of the recipe
      * @return The URL of the image
      */
-    fun updateImage(file: MultipartFile, recipeId:Int) :FileDto?
+    fun updateImage(file: MultipartFile, imageName: String, userId: String) :FileDto?
     {
-        val recipe : PublicRecipe = recipeDao?.findById(recipeId)?.get() ?: return null
+        val auth : FirebaseAuthentication = SecurityContextHolder.getContext().authentication as FirebaseAuthentication
+        if(FirebaseAuth.getInstance().getUser(auth.principal as String).displayName == userId) {
 
-        val localPath = getLocalPathFromImage(recipe.picture)
+            val localPath = "$userId/$imageName"
 
-        val localFile = File(localPath)
+            val localFile = File(localPath)
 
-        if(localFile.exists()) localFile.delete()
+            if (localFile.exists()) localFile.delete()
 
-        localFile.writeBytes(file.bytes)
-        return FileDto("api/$localPath")
+            localFile.writeBytes(file.bytes)
+            return FileDto("api/$localPath")
+        }
+        return null
     }
 
     /**
@@ -100,16 +103,17 @@ class FileService {
      * @param recipeId The id of the recipe
      * @return Empty path
      */
-    fun deleteImage(recipeId: Int) : FileDto?
+    fun deleteImage(imageName: String, userId: String) : FileDto?
     {
-        val recipe : PublicRecipe = recipeDao?.findById(recipeId)?.get() ?: return null
+        val auth : FirebaseAuthentication = SecurityContextHolder.getContext().authentication as FirebaseAuthentication
+        if(FirebaseAuth.getInstance().getUser(auth.principal as String).displayName == userId) {
 
-        val localFile = File(getLocalPathFromImage(recipe.picture))
-        if(localFile.exists()) localFile.delete()
+            val localFile = File("$userId/$imageName")
+            if (localFile.exists()) localFile.delete()
 
-        recipe.picture = ""
-        recipeDao.save(recipe)
-        return FileDto("")
+            return FileDto("")
+        }
+        return null
     }
 
     /**
